+++ kernel/memory/page_allocator.h
@@ -31,9 +31,9 @@ void mark_used(uintptr_t address, size_t pages);
 
 bool page_used(uintptr_t ptr);
 
-void* kalloc_inner(void *page, uint64_t size, uint16_t alignment, uint8_t level, uintptr_t page_va, uintptr_t *next_va, uintptr_t *ttbr);
-void* kalloc(void *page, uint64_t size, uint16_t alignment, uint8_t level);
-void kfree(void* ptr, uint64_t size);
+void* kalloc_inner(void *page, size_t size, uint16_t alignment, uint8_t level, uintptr_t page_va, uintptr_t *next_va, uintptr_t *ttbr);
+void* kalloc(void *page, size_t size, uint16_t alignment, uint8_t level);
+void kfree(void* ptr, size_t size);
 
 int count_pages(uint64_t i1,uint64_t i2);
