+++ shared/data_struct/hashmap.h
@@ -5,8 +5,9 @@
 extern "C" {
 #endif
 
-extern void* malloc(uint64_t size);
-extern void free_sized(void* ptr, uint64_t size);
+//TODO: review allocs & C
+extern void* malloc(size_t size);
+extern void free_sized(void* ptr, size_t size);
 
 typedef uint64_t (*chashmap_hash_fn)(const void* key, uint64_t len);
 typedef int (*chashmap_keyeq_fn)(const void* a, uint64_t alen, const void* b, uint64_t blen);
@@ -23,8 +24,8 @@ typedef struct chashmap {
     chashmap_entry_t** buckets;
     uint64_t capacity;
     uint64_t size;
-    void* (*alloc)(uint64_t size);
-    void (*free)(void* ptr, uint64_t size);
+    void* (*alloc)(size_t size);
+    void (*free)(void* ptr, size_t size);
     chashmap_hash_fn hash_fn;
     chashmap_keyeq_fn keyeq_fn;
     void (*value_dispose)(void* value);
@@ -32,9 +33,9 @@ typedef struct chashmap {
 } chashmap_t;
 
 chashmap_t* chashmap_create(uint64_t initial_capacity);
-chashmap_t* chashmap_create_alloc(uint64_t initial_capacity, void* (*alloc)(uint64_t size),void (*mfree)(void* ptr, size_t size));
+chashmap_t* chashmap_create_alloc(uint64_t initial_capacity, void* (*alloc)(size_t size),void (*mfree)(void* ptr, size_t size));
 void chashmap_destroy(chashmap_t* map);
-void chashmap_set_allocator(chashmap_t* map, void* (*alloc)(uint64_t), void (*dealloc)(void*, uint64_t));
+void chashmap_set_allocator(chashmap_t* map, void* (*alloc)(size_t), void (*dealloc)(void*, size_t));
 void chashmap_set_hash(chashmap_t* map, chashmap_hash_fn hash_fn, chashmap_keyeq_fn keyeq_fn);
 void chashmap_set_value_dispose(chashmap_t* map, void (*dispose_fn)(void*));
 int chashmap_put(chashmap_t* map, const void* key, uint64_t key_len, void* value);
